+++ src/Microsoft.AspNetCore.Razor.Language/Extensions/IDesignTimeDirectiveTargetExtension.cs
@@ -1,9 +1,10 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using Microsoft.AspNetCore.Razor.Language.CodeGeneration;
 using Microsoft.AspNetCore.Razor.Language.Intermediate;
 
-namespace Microsoft.AspNetCore.Razor.Language.CodeGeneration
+namespace Microsoft.AspNetCore.Razor.Language.Extensions
 {
     internal interface IDesignTimeDirectiveTargetExtension : ICodeTargetExtension
     {
