+++ src/Microsoft.AspNetCore.Razor.Language/Extensions/PreallocatedAttributeTargetExtension.cs
@@ -2,10 +2,9 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using Microsoft.AspNetCore.Razor.Language.Intermediate;
-using Microsoft.AspNetCore.Razor.Language.Legacy;
+using Microsoft.AspNetCore.Razor.Language.CodeGeneration;
 
-namespace Microsoft.AspNetCore.Razor.Language.CodeGeneration
+namespace Microsoft.AspNetCore.Razor.Language.Extensions
 {
     internal class PreallocatedAttributeTargetExtension : IPreallocatedAttributeTargetExtension
     {
