+++ src/Microsoft.AspNetCore.Razor.Language/Extensions/SetPreallocatedTagHelperPropertyIntermediateNode.cs
@@ -3,8 +3,9 @@
 
 using System;
 using Microsoft.AspNetCore.Razor.Language.CodeGeneration;
+using Microsoft.AspNetCore.Razor.Language.Intermediate;
 
-namespace Microsoft.AspNetCore.Razor.Language.Intermediate
+namespace Microsoft.AspNetCore.Razor.Language.Extensions
 {
     internal sealed class SetPreallocatedTagHelperPropertyIntermediateNode : ExtensionIntermediateNode
     {
